Initial commit, project LogPulse
This project is a command line utility that can be used to monitor a file,
'/tmp/access.log' by default, and report general statistics and high traffic
events periodically.
Every 10 seconds, general statistics are reported (stdout) for all data
appended to the file during those 10 seconds.
Every 120 seconds the associated data is analysed:

  - whenever the total traffic exceeds a certain threshold, on average, a
    message 'High traffic' is reported.
  - whenever the total traffic drops again below the threshold, on average,
    another message 'Back to normal traffic' is reported.
  - both messages give details about the time when the threshold was crossed.
    The default value for the  threshold is 10.0 requests per second.

The contents of the file are expected to be lines in CLF format:
https://publib.boulder.ibm.com/tividd/td/ITWSA/ITWSA_info45/en_US/HTML/guide/c-logs.html#common